fix: update dependencies version
