+++ pyproject.toml
@@ -3,31 +3,31 @@ name = "fastapi-boilerplate"
 version = "0.1.0"
 requires-python = ">=3.12"
 dependencies = [
-    "aiosqlite>=0.21.0",
-    "alembic>=1.17.2",
+    "aiosqlite>=0.22.1",
+    "alembic>=1.18.4",
     "asyncpg>=0.31.0",
-    "fastapi>=0.115.0",
-    "fastapi-pagination[sqlalchemy]>=0.15.3",
-    "fastapi-users[sqlalchemy]>=15.0.1",
+    "fastapi>=0.129.0",
+    "fastapi-pagination[sqlalchemy]>=0.15.10",
+    "fastapi-users[sqlalchemy]>=15.0.4",
     "httpx-oauth>=0.16.1",
     "loguru>=0.7.3",
-    "pydantic-settings>=2.12.0",
-    "tenacity>=9.0.0",
-    "redis[hiredis]>=7.1.0",
-    "sqlmodel>=0.0.27",
-    "uvicorn[standard]>=0.32.0",
+    "pydantic-settings>=2.13.1",
+    "tenacity>=9.1.4",
+    "redis[hiredis]>=7.2.0",
+    "sqlmodel>=0.0.34",
+    "uvicorn[standard]>=0.41.0",
 ]
 
 [dependency-groups]
 dev = [
-    "fakeredis>=2.32.1",
-    "locust>=2.32.5",
-    "pytest>=8.3.3",
+    "fakeredis>=2.34.0",
+    "locust>=2.43.3",
+    "pytest>=9.0.2",
     "pytest-asyncio>=1.2.0",
     "pytest-cov>=6.1.1",
     "pytest-timeout>=2.4.0",
     "python-dotenv>=1.2.1",
-    "ruff>=0.12.8",
+    "ruff>=0.15.1",
 ]
 
 [build-system]
@@ -42,6 +42,7 @@ packages = ["src"]
 
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
+asyncio_default_fixture_loop_scope = "function"
 # filterwarnings = ["error::DeprecationWarning"]
 timeout = 5
 pythonpath = ["."]
