+++ tests/conftest.py
@@ -7,6 +7,7 @@ from fakeredis import FakeAsyncRedis
 from fastapi import FastAPI
 from httpx import ASGITransport, AsyncClient
 from loguru import logger
+from redis.asyncio.client import Redis as AsyncRedisClient
 from sqlalchemy import text
 from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
 from sqlalchemy.orm import sessionmaker
@@ -19,6 +20,11 @@ from src.session import get_session
 
 env_path = Path(__file__).parent.parent / ".env-example"
 dotenv.load_dotenv(dotenv_path=env_path)
+
+
+class _FakeRedisClientNoDeprecatedParams(AsyncRedisClient):
+    def __init__(self, *, decode_responses: bool = False, **kwargs):
+        super().__init__(decode_responses=decode_responses, **kwargs)
 
 
 @pytest.fixture
@@ -99,12 +105,13 @@ async def test_db():
 
     app.dependency_overrides[get_session] = override_get_session
 
-    yield async_session
-
-    async with engine.begin() as conn:
-        await conn.run_sync(SQLModel.metadata.drop_all)
-
-    app.dependency_overrides.clear()
+    try:
+        yield async_session
+    finally:
+        async with engine.begin() as conn:
+            await conn.run_sync(SQLModel.metadata.drop_all)
+        await engine.dispose()
+        app.dependency_overrides.clear()
 
 
 @pytest.fixture
@@ -117,7 +124,9 @@ async def local_cache():
 @pytest.fixture
 async def redis_cache():
     cache = RedisCache(host="localhost", port=6379, db=1)
-    cache._client = FakeAsyncRedis(decode_responses=True)
+    cache._client = FakeAsyncRedis(
+        decode_responses=True, client_class=_FakeRedisClientNoDeprecatedParams
+    )
     await cache.clear()
     yield cache
     await cache.clear()
